requirements.txt
langchain
openai
python-dotenv
mysql-connector-python
langchain-community
langchain_experimental
langchain_openai
pymysql
streamlit